move ssh node pool functions from sky core to ssh node pools core
