+++ sky/core.py
@@ -1292,41 +1292,6 @@ def local_down() -> None:
                                        is_local=True))
 
 
-@usage_lib.entrypoint
-def ssh_up(infra: Optional[str] = None, cleanup: bool = False) -> None:
-    """Deploys or tears down a Kubernetes cluster on SSH targets.
-
-    Args:
-        infra: Name of the cluster configuration in ssh_node_pools.yaml.
-            If None, the first cluster in the file is used.
-        cleanup: If True, clean up the cluster instead of deploying.
-    """
-    kubernetes_deploy_utils.deploy_ssh_cluster(
-        cleanup=cleanup,
-        infra=infra,
-    )
-
-
-@usage_lib.entrypoint
-def ssh_status(context_name: str) -> Tuple[bool, str]:
-    """Check the status of an SSH Node Pool context.
-
-    Args:
-        context_name: The SSH context name (e.g., 'ssh-my-cluster')
-
-    Returns:
-        Tuple[bool, str]: (is_ready, reason)
-            - is_ready: True if the SSH Node Pool is ready, False otherwise
-            - reason: Explanation of the status
-    """
-    try:
-        is_ready, reason = clouds.SSH.check_single_context(context_name)
-        return is_ready, reason
-    except Exception as e:  # pylint: disable=broad-except
-        return False, ('Failed to check SSH context: '
-                       f'{common_utils.format_exception(e)}')
-
-
 def get_all_contexts() -> List[str]:
     """Get all available contexts from Kubernetes and SSH clouds.
